+++ examples/00_random_field/02_fancier.py
