+++ examples/00_random_field/04_srf_merge.py
