+++ examples/04_kriging/02_pykrige_interface.py
@@ -1,8 +1,14 @@
 """
-PyKrige Interface
-=================
+Interface to PyKrige
+--------------------
 
-Example how to use the PyKrige routines with a GSTools CovModel.
+To use fancier methods like `universal kriging <https://en.wikipedia.org/wiki/Kriging#Universal>`__,
+we provide an interface to `PyKrige <https://github.com/bsmurphy/PyKrige>`__.
+
+You can pass a GSTools Covariance Model to the PyKrige routines as ``variogram_model``.
+
+To demonstrate the general workflow, we compare the ordinary kriging of PyKrige
+with GSTools in 2D:
 """
 import numpy as np
 from gstools import Gaussian, krige
@@ -19,18 +25,27 @@ data = np.array(
         [4.7, 3.8, 1.74],
     ]
 )
+
 # grid definition for output field
 gridx = np.arange(0.0, 5.5, 0.1)
 gridy = np.arange(0.0, 6.5, 0.1)
+
+###############################################################################
+
 # a GSTools based covariance model
 cov_model = Gaussian(
     dim=2, len_scale=1, anis=0.2, angles=-0.5, var=0.5, nugget=0.1
 )
+
+###############################################################################
 # ordinary kriging with pykrige
 OK1 = OrdinaryKriging(data[:, 0], data[:, 1], data[:, 2], cov_model)
 z1, ss1 = OK1.execute("grid", gridx, gridy)
 plt.imshow(z1, origin="lower")
 plt.show()
+
+###############################################################################
+
 # ordinary kriging with gstools for comparison
 OK2 = krige.Ordinary(cov_model, [data[:, 0], data[:, 1]], data[:, 2])
 OK2.structured([gridx, gridy])
