+++ examples/05_conditioned_fields/00_condition_ensemble.py
@@ -1,6 +1,10 @@
 """
-Condition Ensemble
-==================
+Example: Conditioning with Ordinary Kriging
+-------------------------------------------
+
+Here we use ordinary kriging in 1D (for plotting reasons) with 5 given observations/conditions,
+to generate an ensemble of conditioned random fields.
+The estimated mean can be accessed by ``srf.mean``.
 """
 import numpy as np
 from gstools import Gaussian, SRF
@@ -10,10 +14,15 @@ import matplotlib.pyplot as plt
 cond_pos = [0.3, 1.9, 1.1, 3.3, 4.7]
 cond_val = [0.47, 0.56, 0.74, 1.47, 1.74]
 gridx = np.linspace(0.0, 15.0, 151)
+
+###############################################################################
+
 # spatial random field class
 model = Gaussian(dim=1, var=0.5, len_scale=2)
 srf = SRF(model)
 srf.set_condition(cond_pos, cond_val, "ordinary")
+
+###############################################################################
 fields = []
 for i in range(100):
     if i % 10 == 0:
@@ -27,3 +36,7 @@ plt.plot(gridx, srf.krige_field, linestyle="dashed", label="kriged field")
 plt.scatter(cond_pos, cond_val, color="k", zorder=10, label="Conditions")
 plt.legend()
 plt.show()
+
+###############################################################################
+# As you can see, the kriging field coincides with the ensemble mean of the
+# conditioned random fields and the estimated mean is the mean of the far-field.
