+++ examples/06_transformations/01_binary.py
@@ -1,6 +1,10 @@
 """
-Transform 1
-===========
+binary fields
+-------------
+
+Here we transform a field to a binary field with only two values.
+The dividing value is the mean by default and the upper and lower values
+are derived to preserve the variance.
 """
 from gstools import SRF, Gaussian
 from gstools import transform as tf
@@ -10,5 +14,5 @@ x = y = range(100)
 model = Gaussian(dim=2, var=1, len_scale=10)
 srf = SRF(model, seed=20170519)
 srf.structured([x, y])
-tf.normal_to_lognormal(srf)
+tf.binary(srf)
 srf.plot()
