+++ examples/06_transformations/README.rst
@@ -1,0 +1,43 @@
+Tutorial 7: Field transformations
+=================================
+
+The generated fields of gstools are ordinary Gaussian random fields.
+In application there are several transformations to describe real world
+problems in an appropriate manner.
+
+GStools provides a submodule :py:mod:`gstools.transform` with a range of
+common transformations:
+
+.. currentmodule:: gstools.transform
+
+.. autosummary::
+   binary
+   boxcox
+   zinnharvey
+   normal_force_moments
+   normal_to_lognormal
+   normal_to_uniform
+   normal_to_arcsin
+   normal_to_uquad
+
+
+Implementation
+--------------
+
+All the transformations take a field class, that holds a generated field,
+as input and will manipulate this field inplace.
+
+Simply import the transform submodule and apply a transformation to the srf class:
+
+.. code-block:: python
+
+    from gstools import transform as tf
+    ...
+    tf.normal_to_lognormal(srf)
+
+
+In the following we will start from a simple random field following a Gaussian covariance:
+
+.. image:: pics/07_00_std.png
+   :width: 600px
+   :align: center
