--- examples/10_simple_kriging.py
@@ -1,19 +1,0 @@
-"""
-Simple Kriging
-==============
-"""
-import numpy as np
-from gstools import Gaussian, krige
-
-# condtions
-cond_pos = [0.3, 1.9, 1.1, 3.3, 4.7]
-cond_val = [0.47, 0.56, 0.74, 1.47, 1.74]
-# resulting grid
-gridx = np.linspace(0.0, 15.0, 151)
-# spatial random field class
-model = Gaussian(dim=1, var=0.5, len_scale=2)
-krig = krige.Simple(model, mean=1, cond_pos=cond_pos, cond_val=cond_val)
-krig(gridx)
-ax = krig.plot()
-ax.scatter(cond_pos, cond_val, color="k", zorder=10, label="Conditions")
-ax.legend()
