--- examples/15_transform_02.py
@@ -1,14 +1,0 @@
-"""
-Transform 2
-===========
-"""
-from gstools import SRF, Gaussian
-from gstools import transform as tf
-
-# structured field with a size of 100x100 and a grid-size of 1x1
-x = y = range(100)
-model = Gaussian(dim=2, var=1, len_scale=10)
-srf = SRF(model, seed=20170519)
-srf.structured([x, y])
-tf.binary(srf)
-srf.plot()
